Non-interavtive use (v2)
Alternative to #189 for non-interactive us (e.g. in cron jobs or via window-manager bindings).

Examples:

    # Play/pause.
    castnow --command space --exit

    # Louder.
    castnow --command up --exit

    # Quieter.
    castnow --command down --exit

The choice of `--command` is intended to be analogous to `-c` for shells.

On the positive side, this is probably better that #189; it's similar to @hemanth's `--flags` suggestion there.

On the negative side, it still has the oddity of using key names for
command names.  I'm not sure what to do about that.  On the one hand, it
certainly is a bit odd; on the other, I'd prefer not to have to
introduce an entire new vocabulary (and have to document and maintain that
vocabulary).

In summary, this seems to be a reasonable compromise.